requirements.txt
Django==5.0.7
psycopg2==2.9.9
django-crispy-forms==2.2
djangorestframework==3.15.2